Enable the use of different links for reference / to apply forces/torques in propulsion and aerodynamics plugin...
Became necessary for tilt-wing simulation where forces/torques on actuated wing are considerable. Wing joint control cannot be made stiff enough and degrades. As remedy, we apply equivalent forces/torques directly on the baselink (fuselage) instead of the wing - and use the wing link (unloaded) only as reference to calculate inflow conditions and all the necessary transforms...

Backwards compatible with xacros using the earlier implementations of these plugins where indicated link was used for reference *and* to apply forces to.